Impl GrafoT for Arbol
